@@ -1,28 +1,43 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using ReactiveUI;
+using System.ComponentModel;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
 
-namespace VigilantMonitor
+namespace VigilantMonitor;
+
+public partial class MainWindow
 {
-    /// <summary>
-    /// Interaction logic for MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow : Window
+    public MainWindow()
     {
-        public MainWindow()
+        InitializeComponent();
+
+        DataContext = ViewModel = RxApp.SuspensionHost.GetAppState<MainViewModel>();
+
+        Validation.AddErrorHandler(this, MainWindow_OnValidationError);
+
+        this.WhenActivated(disposables =>
         {
-            InitializeComponent();
-        }
+            //this.Bind(ViewModel,
+            //        vm => vm.MinDuration,
+            //        v => v.MinDurationText.Text)
+            //    .DisposeWith(disposables);
+
+            //this.BindValidation(ViewModel,
+            //    vm => vm.MinDuration,
+            //    v => v.MinDurationValidationText.Text)
+            //    .DisposeWith(disposables);
+        });
+    }
+
+    private void MainWindow_OnValidationError(object? sender, ValidationErrorEventArgs e)
+    {
+        if (e.Error.RuleInError.ValidationStep == ValidationStep.ConvertedProposedValue)
+            ViewModel!.HasConversionError = e.Action == ValidationErrorEventAction.Added;
+    }
+
+    private void MainWindow_OnClosing(object? sender, CancelEventArgs e)
+    {
+        Visibility = Visibility.Hidden;
+        e.Cancel = true;
     }
 }
